requirements.txt
# ML and AI dependencies
sentence-transformers>=2.2.0,<3.0.0
torch>=1.11.0,<3.0.0
numpy>=1.21.0,<1.27.0

# Database connectivity
cassandra-driver>=3.28.0,<4.0.0
astrapy>=2.0.0,<3.0.0

# Data processing
pandas>=1.5.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0

# Language detection
langdetect>=1.0.0,<2.0.0

# OpenAI integration (optional)
openai>=1.0.0,<2.0.0

# Web framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0 